fixes, fixes, fixes...

+++ src/main/java/site/geni/FarLands/mixins/client/block/RepeaterBlockMixin.java
@@ -55,8 +55,8 @@ public abstract class RepeaterBlockMixin extends HorizontalFacingBlockMixin {
 			}
 			float_1 /= 16.0F;
 
-			double offsetX = (double) (float_1 * (float) direction.getOffsetX());
-			double offsetZ = (double) (float_1 * (float) direction.getOffsetZ());
+			double offsetX = float_1 * (float) direction.getOffsetX();
+			double offsetZ = float_1 * (float) direction.getOffsetZ();
 
 			world.addParticle(DustParticleEffect.RED, x + offsetX, y, z + offsetZ, 0.0D, 0.0D, 0.0D);
